Refactor Client alerts.
- Extract anomalous store inside /components namespace.
