+++ app/javascript/alerts/components/alerts/alerts.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Alert from './alert';
-import { alertListType } from './entities';
+import Alert from '../../../components/alert';
+import { alertListType } from '../../entities';
 
 const Alerts = ({ alerts, dismissAlert }) => (
   alerts.map(alert => (<Alert key={alert.id} alert={alert} dismissAlert={dismissAlert} />))
