+++ app/javascript/alerts/components/alerts/alerts.test.jsx
