+++ app/javascript/alerts/entities.jsx
