+++ app/javascript/alerts/index.js
@@ -1,0 +1,3 @@
+export { default as Alerts } from './components/alerts';
+
+export { default as reducer } from './store';
