+++ app/javascript/alerts/store/index.js
@@ -1,0 +1,3 @@
+import reducer from './reducer';
+
+export default reducer;
