+++ app/javascript/alerts/store/initialState.js
