+++ app/javascript/alerts/store/initialState.test.js
