+++ app/javascript/components/alert/dismissButton.jsx
