+++ app/javascript/mechanics/actions/store/deleteAction/alerts.js
@@ -1,4 +1,4 @@
-import { addAlert } from '../../../../components/alerts/store/actions';
+import { addAlert } from '../../../../alerts/store/actions';
 import { generateFingerprintUuid } from '../../../../utils/uuid';
 
 const failureAlert = {
